refactor(core): move queue into queuing module
